Change OptionView location
