+++ sdk/src/main/kotlin/io/rover/sdk/data/domain/Experience.kt
@@ -118,7 +118,6 @@ interface Block {
     companion object
 }
 
-
 data class ImageBlockOption(
     val text: String,
     val image: PollImage
